+++ packages/peregrine/lib/talons/MiniCart/useKebab.js
@@ -1,0 +1,27 @@
+import { useCallback, useRef, useState } from 'react';
+import { useEventListener } from '@magento/peregrine';
+
+export const useKebab = () => {
+    const kebabRef = useRef(null);
+    const [isOpen, setIsOpen] = useState(false);
+
+    const handleKebabClick = useCallback(() => {
+        setIsOpen(!isOpen);
+    }, [isOpen]);
+
+    const handleOutsideKebabClick = useCallback(event => {
+        // Ensure we're truly outside of the kebab.
+        if (!kebabRef.current.contains(event.target)) {
+            setIsOpen(false);
+        }
+    }, []);
+
+    useEventListener(document, 'mousedown', handleOutsideKebabClick);
+    useEventListener(document, 'touchend', handleOutsideKebabClick);
+
+    return {
+        handleKebabClick,
+        isOpen,
+        kebabRef
+    };
+};
