+++ packages/venia-ui/lib/components/MiniCart/kebab.js
@@ -1,35 +1,18 @@
-import React, { useCallback, useRef, useState } from 'react';
+import React from 'react';
 import { node, shape, string } from 'prop-types';
 import { MoreVertical as MoreVerticalIcon } from 'react-feather';
-
-import { useEventListener } from '@magento/peregrine';
 
 import { mergeClasses } from '../../classify';
 import Icon from '../Icon';
 
 import defaultClasses from './kebab.css';
+import { useKebab } from '@magento/peregrine/lib/talons/MiniCart/useKebab';
 
 const Kebab = props => {
+    const { handleKebabClick, isOpen, kebabRef } = useKebab();
     const { children } = props;
-
-    const kebabRef = useRef(null);
-    const [isOpen, setIsOpen] = useState(false);
-
     const classes = mergeClasses(defaultClasses, props.classes);
     const toggleClass = isOpen ? classes.dropdown_active : classes.dropdown;
-
-    const handleKebabClick = useCallback(() => {
-        setIsOpen(!isOpen);
-    }, [isOpen]);
-    const handleOutsideKebabClick = useCallback(event => {
-        // Ensure we're truly outside of the kebab.
-        if (!kebabRef.current.contains(event.target)) {
-            setIsOpen(false);
-        }
-    }, []);
-
-    useEventListener(document, 'mousedown', handleOutsideKebabClick);
-    useEventListener(document, 'touchend', handleOutsideKebabClick);
 
     return (
         <div className={classes.root}>
